refactor Seperate all classes
